Inteligencia Artificial/(05) Integracao com APIs de LLMs/(06) Python - Primeiros passos com a API da Anthropic (Claude)/main.py
import os
from dotenv import load_dotenv

load_dotenv()

# from providers import openai
from providers import anthropic
# from providers import xai_sdk
# from providers import google as gai
# from providers import groq
# from providers import ollama

# print(openai.test())
print(anthropic.test())