Inteligencia Artificial/(05) Integracao com APIs de LLMs/(06) Python - Primeiros passos com a API da Anthropic (Claude)/requirements.txt
python-dotenv

openai
anthropic
google-generativeai
xai-sdk
groq
ollama